11-3 检索服务响应对象的定义

+++ imooc-ad-service/ad-common/src/main/java/com/imooc/ad/dump/table/AdCreativeTable.java
@@ -12,8 +12,8 @@ public class AdCreativeTable {
   private String name;
   private Integer type;
   private Integer materialType;
-  private Long height;
-  private Long width;
+  private Integer height;
+  private Integer width;
   private Integer auditStatus;
   private String adUrl;
 }
